Modifications to logger in kernel.cpp, added more logging messages, remove all I/O system calls from signal handlers, solved a race condition between SIGCHLD and checking the requests msg queue

+++ kernel.cpp
@@ -1,5 +1,3 @@
-#include "logger.h"
-
 #include <iostream>
 #include <signal.h>
 #include <unistd.h>
@@ -9,16 +7,17 @@
 #include <sys/msg.h>
 #include <wait.h>
 #include <queue>
+#include "logger.h"
 
-#define PROCESSES_COUNT 1
+#define PROCESSES_COUNT 2
 #define PROCESS_EXECUTABLE_FILE "process"
 #define DISK_EXECUTABLE_FILE "disk"
 #define PROCESS_INPUT_FILE_NAME "processInput"
+#define LOG_FILE_NAME "log.txt"
 #define ADD_REQ 1
 #define DELETE_REQ 2
 #define MSG_SIZE 30
-#define MSG_QUEUE_FLAGS 0666 
-
+#define MSG_QUEUE_FLAGS 0666
 
 using namespace std;
 
@@ -27,9 +26,10 @@ struct msgbuff {
     char data[MSG_SIZE];
 };
 
+const int DATA_SIZE = sizeof(msgbuff) - sizeof(long);
 int currentTime, runningProcesses;
-bool isDiskAvailable;
-logger logfile;
+bool isDiskAvailable, processesRunning;
+Logger logger;
 
 void setUpSignalHandlers();
 pid_t createDisk(string diskFilePath, key_t toDiskQueueID, key_t fromDiskQueueID);
@@ -41,12 +41,14 @@ void sendDiskOperation(queue<msgbuff>& requests, key_t toDiskQueueID,
 void createMessageQueues(key_t& requestsQueueID, key_t& toDiskQueueID, key_t& fromDiskQueueID);
 void terminateMessageQueues(key_t& requestsQueueID, key_t& toDiskQueueID, key_t& fromDiskQueueID);
 void terminateDisk(pid_t diskPID);
+void logTerminating();
 
 
 int main() {
-    logfile.openFile("logfile.txt");
+    logger.openFile(LOG_FILE_NAME);
     currentTime = 1;
     isDiskAvailable = true;
+    processesRunning = true;
     key_t requestsQueueID, toDiskQueueID, fromDiskQueueID;
     queue<msgbuff> requests;
     createMessageQueues(requestsQueueID, toDiskQueueID, fromDiskQueueID);
@@ -54,21 +56,20 @@ int main() {
     int diskPID = createDisk(DISK_EXECUTABLE_FILE, toDiskQueueID, fromDiskQueueID);
     createProcesses(PROCESS_EXECUTABLE_FILE, PROCESS_INPUT_FILE_NAME, PROCESSES_COUNT, requestsQueueID);
     alarm(1);
-    while(runningProcesses > 0 || !requests.empty() || !isDiskAvailable) {
+    while(processesRunning || !requests.empty() || !isDiskAvailable) {
         getRequests(requests, requestsQueueID);
         sendDiskOperation(requests, toDiskQueueID, fromDiskQueueID, diskPID);
     }
     terminateDisk(diskPID);
     terminateMessageQueues(requestsQueueID, toDiskQueueID, fromDiskQueueID);
-    logfile.closeFile();
+    logTerminating();
+    logger.closeFile();
 }
 
 /************** Termination Handling **************/
 
 void terminateDisk(pid_t diskPID) {
-    logfile.writeToFile("All processes exited, terminating disk...");
     kill(diskPID, SIGKILL);
-    logfile.writeToFile("Exiting...");
 }
 
 void terminateMessageQueues(key_t& requestsQueueID, key_t& toDiskQueueID, key_t& fromDiskQueueID) {
@@ -77,10 +78,44 @@ void terminateMessageQueues(key_t& requestsQueueID, key_t& toDiskQueueID, key_t&
     msgctl(fromDiskQueueID, IPC_RMID, nullptr);
 }
 
+/************** Logging **************/
+
+void logRequestReceived(const msgbuff& request) {
+    string logData = "Request received, type: ";
+    logData += request.mtype == ADD_REQ ? "ADD" : "DEL";
+    logData += ", data: ";
+    logData += request.data;
+    logger.log(logData);
+}
+
+void logFreeSlotsChecked() {
+    string logData = "Disk is being checked for free slots";
+    logger.log(logData);
+}
+
+void logCheckDiskResponse(const msgbuff& response) {
+    string logData = "Disk has ";
+    logData += response.data[0];
+    logData += " free slot(s)";
+    logger.log(logData);
+}
+
+void logRequestSentToDisk(const msgbuff& request) {
+    string logData = "Request sent to disk, type: ";
+    logData += request.mtype == ADD_REQ ? "ADD" : "DEL";
+    logData += ", data: ";
+    logData += request.data;
+    logger.log(logData);
+}
+
+void logTerminating() {
+    logger.log("All processes exited and all requests are served, exiting..");
+}
+
 /************** Requests Handling **************/
 
 bool checkRequest(key_t requestsQueueID, msgbuff& request) {
-    int status = msgrcv(requestsQueueID, &request, MSG_SIZE, 0, IPC_NOWAIT);
+    int status = msgrcv(requestsQueueID, &request, DATA_SIZE, 0, IPC_NOWAIT);
     if(status != -1) {
         return true;
     }
@@ -92,7 +127,7 @@ bool checkRequest(key_t requestsQueueID, msgbuff& request) {
 void getRequests(queue<msgbuff>& requests, key_t requestsQueueID) {
     msgbuff request;
     while(checkRequest(requestsQueueID, request)) {
-        logfile.writeToFile("Request received, type: " + std::to_string(request.mtype) + " data: " + request.data);
+        logRequestReceived(request);
         requests.push(request);
     }
 }
@@ -107,14 +142,17 @@ void clearAddRequests(queue<msgbuff>& requests) {
 
 bool checkDiskFreeSlots(key_t fromDiskQueueID, pid_t diskPID) {
     kill(diskPID, SIGUSR1);
+    logFreeSlotsChecked();
     msgbuff diskResponse;
-    int status = msgrcv(fromDiskQueueID, &diskResponse, MSG_SIZE, 0, 0);
+    int status = msgrcv(fromDiskQueueID, &diskResponse, DATA_SIZE, 0, 0);
     int freeSlots = int(diskResponse.data[0]) - int('0');
+    logCheckDiskResponse(diskResponse);
     return freeSlots > 0;
 }
 
 void sendToDisk(key_t toDiskQueueID, msgbuff* msg) {
-    msgsnd(toDiskQueueID, msg, MSG_SIZE, IPC_NOWAIT);
+    logRequestSentToDisk(*msg);
+    msgsnd(toDiskQueueID, msg, DATA_SIZE, IPC_NOWAIT);
 }
 
 void sendDiskOperation(queue<msgbuff>& requests, 
@@ -156,13 +194,18 @@ void SIGUSR2_handler(int signum) {
 
 void SIGCHLD_handler(int signum) {
     int status;
-    while(waitpid(-1, &status, WNOHANG) > 0)
+    while(waitpid(-1, &status, WNOHANG) > 0) {
         runningProcesses--;
+    }
 }
 
 void SIGALRM_handler(int signum) {
-    alarm(1);
-    killpg(getpgrp(), SIGUSR2);
+    if(runningProcesses == 0)
+        processesRunning = false;
+    else {
+        alarm(1);
+        killpg(getpgrp(), SIGUSR2);
+    }
 }
 
 void setUpSignalHandlers() {
